Reverting a refactor of status and error
I moved status and error out of player and player item into a
protocol which I think was a good call, but combined the two
into one observable stream. This meant an error ended your
status stream. I think this was a mistake, and I am reverting
it while I think on the best approach to this.